+++ tests/StackExchange.Redis.Tests/KeyPrefixedTests.cs
@@ -941,6 +941,21 @@ namespace StackExchange.Redis.Tests
         }
 
         [Fact]
+        public async Task StreamNegativeAcknowledgeAsync_1()
+        {
+            await prefixed.StreamNegativeAcknowledgeAsync("key", "group", "consumer", StreamNackMode.Fail, "0-0", CommandFlags.None);
+            await mock.Received().StreamNegativeAcknowledgeAsync("prefix:key", "group", "consumer", StreamNackMode.Fail, "0-0", CommandFlags.None);
+        }
+
+        [Fact]
+        public async Task StreamNegativeAcknowledgeAsync_2()
+        {
+            var messageIds = new RedisValue[] { "0-0", "0-1", "0-2" };
+            await prefixed.StreamNegativeAcknowledgeAsync("key", "group", "consumer", StreamNackMode.Fail, messageIds, CommandFlags.None);
+            await mock.Received().StreamNegativeAcknowledgeAsync("prefix:key", "group", "consumer", StreamNackMode.Fail, messageIds, CommandFlags.None);
+        }
+
+        [Fact]
         public async Task StreamAddAsync_1()
         {
             await prefixed.StreamAddAsync("key", "field1", "value1", "*", 1000, true, CommandFlags.None);